Adds list actions
