+++ src/components/ConfigEditor/index.module.css
@@ -3,12 +3,13 @@
   border-radius: 10px;
   box-shadow: 0.3rem 0.3rem 0.6rem rgba(var(--theme-bg-1-rgb), 0.4) -0.2rem -0.2rem
     0.5rem rgba(var(--theme-bg-3-rgb), 0.3);
+  width: 100%;
+  height: 100%;
 }
 
 @media (max-width: 600px) {
   .modal {
     border-radius: 0;
-    width: 100%;
   }
 }
