+++ src/index.css
@@ -6,12 +6,14 @@
   --theme-color-disabled: #555;
   --theme-highlight-1: gold;
   --theme-warning-1: #f44336;
+  --theme-success-1: #87cc58;
 
   --theme-bg-1-rgb: 0, 0, 0;
   --theme-bg-2-rgb: 21, 21, 21;
   --theme-bg-3-rgb: 45, 45, 45;
   --theme-highlight-1-rgb: 255, 215, 0;
   --theme-warning-1-rgb: 244, 67, 54;
+  --theme-success-1-rgb: 135, 204, 88;
 
   --ease-1: cubic-bezier(0.075, 0.82, 0.165, 1);
 }
